+++ test/controllers/donations_controller_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+describe DonationsController do
+  describe '#index' do
+    it 'ページを表示する' do
+      get donations_path
+
+      assert_response :success
+    end
+  end
+end
