+++ test/controllers/users/registrations_controller_test.rb
